Added transactions for bulk raffle ticket save and winner update. Sage Role assignment moved to discord/actions and changed role names to an enum.
