+++ src/challenges.ts
@@ -4,7 +4,6 @@ import {
   ChallengeDifficulty,
   Region,
 } from '../src/database/models';
-import { Guild, GuildMember } from 'discord.js';
 
 type ChallengeCache = {
   challenges: Challenge[];
@@ -290,47 +289,5 @@ export function getDifficultyName(difficulty: ChallengeDifficulty): string {
       return 'Grandmaster';
     default:
       return 'Unknown';
-  }
-}
-
-/**
- * Assigns the user a role based on their difficulty level.
- * @param guild - The Discord guild where the roles are located.
- * @param discordId - The user's Discord ID.
- * @param difficulty - The difficulty enum for the user.
- */
-export async function assignSageRole(
-  guild: Guild,
-  discordId: string,
-  difficulty: ChallengeDifficulty,
-): Promise<void> {
-  const member: GuildMember | null = await guild.members.fetch(discordId);
-
-  if (!member) {
-    console.error(
-      `User with ID ${discordId} not found. Could not assign Sage's role`,
-    );
-    return;
-  }
-
-  // Determine role based on difficulty
-  let roleName: string | null = null;
-
-  if (difficulty === ChallengeDifficulty.MASTER) {
-    roleName = "Sage's Master";
-  } else if (difficulty === ChallengeDifficulty.GRANDMASTER) {
-    roleName = "Sage's Grandmaster";
-  }
-
-  // Assign the role if applicable
-  if (roleName) {
-    const role = guild.roles.cache.find((r) => r.name === roleName);
-
-    if (!role) {
-      console.error(`Role "${roleName}" not found.`);
-      return;
-    }
-
-    await member.roles.add(role);
   }
 }
